proyecto_db_django/requirements.txt

django
djangorestframework
mysqlclient
psycopg2
djongo
